+++ R/layout_stress.R
@@ -13,6 +13,7 @@
 #'
 #' The layout_igraph_* function should not be used directly. It is only used as an argument for plotting with 'igraph'.
 #' 'ggraph' natively supports the layout.
+#' @seealso [layout_stress3D]
 #' @return matrix of xy coordinates
 #' @references Gansner, E. R., Koren, Y., & North, S. (2004). Graph drawing by stress majorization. *In International Symposium on Graph Drawing* (pp. 239-250). Springer, Berlin, Heidelberg.
 #' @examples
@@ -115,9 +116,119 @@ layout_with_stress <- function(g,weights = NA, iter = 500,tol = 0.0001,mds = TRU
   x
 }
 
-#-------------------------------------------------------------------------------
-# focal layout
-#-------------------------------------------------------------------------------
+#------------------------------------------------------------------------------#
+#------------------------------------------------------------------------------#
+
+#' stress majorization layout in 3D
+#'
+#' @name layout_stress3D
+#' @rdname layout_stress3D
+#' @description force-directed graph layout based on stress majorization in 3D.
+#' @param g igraph object
+#' @param weights possibly a numeric vector with edge weights. If this is NULL and the graph has a weight edge attribute, then the attribute is used. If this is NA then no weights are used (even if the graph has a weight attribute). By default, weights are ignored. See details for more.
+#' @param iter number of iterations during stress optimization
+#' @param tol stopping criterion for stress optimization
+#' @param mds should an MDS layout be used as initial layout (default: TRUE)
+#' @param bbox constrain dimension of output. Only relevant to determine the placement of disconnected graphs
+#' @details Be careful when using weights. In most cases, the inverse of the edge weights should be used to ensure that the endpoints of an edges with higher weights are closer together (weights=1/E(g)$weight).
+#'
+#' @return matrix of xyz coordinates
+#' @seealso [layout_stress]
+#' @references Gansner, E. R., Koren, Y., & North, S. (2004). Graph drawing by stress majorization. *In International Symposium on Graph Drawing* (pp. 239-250). Springer, Berlin, Heidelberg.
+#' @export
+layout_with_stress3D <- function(g,weights = NA, iter = 500,tol = 0.0001,mds = TRUE,bbox = 30){
+  if (!igraph::is_igraph(g)) {
+    stop("Not a graph object")
+  }
+
+  set.seed(42)  #stress is deterministic and produces same result up to translation. This keeps the layout fixed
+
+  comps <- igraph::components(g,"weak")
+  if(comps$no>1){
+    lg <- list()
+    node_order <- c()
+    for (i in 1:comps$no){
+      sg <- igraph::induced_subgraph(g,comps$membership==i)
+      n <- igraph::vcount(sg)
+      node_order <- c(node_order,which(comps$membership==i))
+      if(n==1){
+        lg[[i]] <- matrix(c(0,0,0),1,3,byrow = TRUE)
+        next()
+      }
+      if(n==2){
+        lg[[i]] <- matrix(c(0,0,0,1,0,0),2,3,byrow = TRUE)
+        next()
+      }
+
+      D <- igraph::distances(sg,weights = weights)
+      W <- 1/D^2
+      diag(W) <- 0
+      if(!mds){
+        xinit <- matrix(stats::runif(n*3,0,1),n,3)
+      } else{
+        rmat <- matrix(stats::runif(n*3,-0.1,0.1),n,3)
+        if(igraph::vcount(sg)<=100){
+          xinit <- igraph::layout_with_mds(sg,dim = 3) + rmat
+        } else{
+          n <- igraph::vcount(g)
+          pivs <- sample(1:n,100)
+          D1 <- D[,pivs]
+          cmean <- colMeans(D1^2)
+          rmean <- rowMeans(D1^2)
+          Dmat <- D1^2-outer(rmean,cmean, function(x,y) x+y)+mean(D1^2)
+          sl2 <- svd(Dmat)
+          rmat <- matrix(stats::runif(n*3,-0.1,0.1),n,3)
+          xinit <- (Dmat%*%sl2$v[,1:3]) + rmat
+        }
+      }
+      lg[[i]] <- stress_major3D(xinit,W,D,iter,tol)
+    }
+    lg <- lapply(lg,mv_to_null)
+    p <- order(comps$csize)
+    curx <- 0
+    cury <- 0
+    maxy <- 0
+    for(comp in p){
+      if(curx+max(lg[[comp]][,1])>bbox){
+        curx <- 0
+        cury <- maxy+1
+      }
+      lg[[comp]][,1] <- lg[[comp]][,1] + curx
+      lg[[comp]][,2] <- lg[[comp]][,2] + cury
+      curx <- max(lg[[comp]][,1])+1
+      maxy <- max(c(maxy,max(lg[[comp]][,2])))
+    }
+    x <- do.call("rbind",lg)
+    x <- x[order(node_order),]
+
+  } else{
+    if(igraph::vcount(g)==1){
+      x <- matrix(c(0,0),1,2)
+    } else{
+      D <- igraph::distances(g,weights = weights)
+      W <- 1/D^2
+      diag(W) <- 0
+      n <- igraph::vcount(g)
+      if(!mds){
+        xinit <- matrix(stats::runif(n*2,0,1),n,2)
+      } else{
+        rmat <- matrix(stats::runif(n*2,-0.1,0.1),n,2)
+        if(igraph::vcount(g)<=100){
+          xinit <- igraph::layout_with_mds(g) + rmat
+        } else{
+          xinit <- layout_with_pmds(g,D = D[,sample(1:(igraph::vcount(g)),100)]) + rmat
+        }
+
+      }
+      x <- stress_major(xinit,W,D,iter,tol)
+    }
+  }
+  x
+}
+
+
+#------------------------------------------------------------------------------#
+#------------------------------------------------------------------------------#
 #' radial focus layout
 #'
 #' @description arrange nodes in concentric circles around a focal node according to their distance from the focus.
@@ -177,9 +288,8 @@ layout_with_focus <- function(g,v,weights = NA,iter = 500,tol = 0.0001){
   return(list(xy=x,distance=D[,v]))
 }
 
-#-------------------------------------------------------------------------------
-# centrality layout
-#-------------------------------------------------------------------------------
+#------------------------------------------------------------------------------#
+#------------------------------------------------------------------------------#
 #' radial centrality layout
 #'
 #' @description arranges nodes in concentric circles according to a centrality index.
@@ -258,7 +368,6 @@ layout_with_centrality <- function(g,cent,scale = TRUE,iter = 500,tol = 0.0001,t
 }
 
 #------------------------------------------------------------------------------#
-# constrained stress ----
 #------------------------------------------------------------------------------#
 #' constrained stress layout
 #'
@@ -276,6 +385,7 @@ layout_with_centrality <- function(g,cent,scale = TRUE,iter = 500,tol = 0.0001,t
 #'
 #' The layout_igraph_* function should not be used directly. It is only used as an argument for plotting with 'igraph'.
 #' 'ggraph' natively supports the layout.
+#' @seealso [layout_constrained_stress3D]
 #' @return matrix of xy coordinates
 #' @references Gansner, E. R., Koren, Y., & North, S. (2004). Graph drawing by stress majorization. *In International Symposium on Graph Drawing* (pp. 239-250). Springer, Berlin, Heidelberg.
 #' @export
@@ -320,10 +430,75 @@ layout_with_constrained_stress <- function(g,coord,fixdim="x",weights = NA,
   x
 }
 
+#' constrained stress layout in 3D
+#'
+#' @name layout_constrained_stress3D
+#' @description force-directed graph layout based on stress majorization with variable constrained in 3D
+#' @param g igraph object
+#' @param coord numeric vector. fixed coordinates for dimension specified in `fixdim`.
+#' @param fixdim string. which dimension should be fixed. Either "x", "y" or "z".
+#' @param weights possibly a numeric vector with edge weights. If this is NULL and the graph has a weight edge attribute, then the attribute is used. If this is NA then no weights are used (even if the graph has a weight attribute). By default, weights are ignored. See details for more.
+#' @param iter number of iterations during stress optimization
+#' @param tol stopping criterion for stress optimization
+#' @param mds should an MDS layout be used as initial layout (default: TRUE)
+#' @param bbox constrain dimension of output. Only relevant to determine the placement of disconnected graphs
+#' @details Be careful when using weights. In most cases, the inverse of the edge weights should be used to ensure that the endpoints of an edges with higher weights are closer together (weights=1/E(g)$weight).
+#'
+#' This function does not come with direct support for igraph or ggraph.
+#'
+#' @seealso [layout_constrained_stress]
+#' @return matrix of xyz coordinates
+#' @references Gansner, E. R., Koren, Y., & North, S. (2004). Graph drawing by stress majorization. *In International Symposium on Graph Drawing* (pp. 239-250). Springer, Berlin, Heidelberg.
+#' @export
+layout_with_constrained_stress3D <- function(g,coord,fixdim="x",weights = NA,
+                                             iter = 500,tol = 0.0001,mds = TRUE,bbox = 30){
+  if (!igraph::is_igraph(g)) {
+    stop("Not a graph object")
+  }
+  set.seed(42)
+  fixdim <- match.arg(fixdim,c("x","y","z"))
+  fixdim <- ifelse(fixdim=="x",1,ifelse(fixdim=="y",2,3))
 
-#-------------------------------------------------------------------------------
-# helper functions
-#-------------------------------------------------------------------------------
+  if(missing(coord)){
+    stop('"coord" is missing with no default.')
+  }
+  comps <- igraph::components(g,"weak")
+  if(comps$no>1){
+    stop('g must be connected')
+  }
+
+  if(igraph::vcount(g)==1){
+    x <- matrix(c(0,0,0),1,3)
+  } else{
+    D <- igraph::distances(g,weights = weights)
+    W <- 1/D^2
+    diag(W) <- 0
+    n <- igraph::vcount(g)
+    if(!mds){
+      xinit <- matrix(stats::runif(n*3,0,1),n,3)
+      xinit[,fixdim] <- coord
+    } else{
+      n <- igraph::vcount(g)
+      pivs <- sample(1:n,min(c(50,n)))
+      D1 <- D[,pivs]
+      cmean <- colMeans(D1^2)
+      rmean <- rowMeans(D1^2)
+      Dmat <- D1^2-outer(rmean,cmean, function(x,y) x+y)+mean(D1^2)
+      sl2 <- svd(Dmat)
+      rmat <- matrix(stats::runif(n*3,-0.1,0.1),n,3)
+      xinit <- (Dmat%*%sl2$v[,1:3]) + rmat
+      row.names(xinit) <- NULL
+
+      xinit[,fixdim] <- coord
+    }
+    x <- constrained_stress_major3D(xinit,fixdim,W,D,iter,tol)
+  }
+  x
+}
+
+#-------------------------------------------------------------------------------#
+# helper functions ----
+#-------------------------------------------------------------------------------#
 
 get_bbox <- function(xy){
   lbottom <- c(min(xy[,1]),min(xy[,2]))