+++ src/RcppExports.cpp
@@ -35,6 +35,35 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// constrained_stress3D
+double constrained_stress3D(NumericMatrix x, NumericMatrix W, NumericMatrix D);
+RcppExport SEXP _graphlayouts_constrained_stress3D(SEXP xSEXP, SEXP WSEXP, SEXP DSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type W(WSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type D(DSEXP);
+    rcpp_result_gen = Rcpp::wrap(constrained_stress3D(x, W, D));
+    return rcpp_result_gen;
+END_RCPP
+}
+// constrained_stress_major3D
+NumericMatrix constrained_stress_major3D(NumericMatrix y, int dim, NumericMatrix W, NumericMatrix D, int iter, double tol);
+RcppExport SEXP _graphlayouts_constrained_stress_major3D(SEXP ySEXP, SEXP dimSEXP, SEXP WSEXP, SEXP DSEXP, SEXP iterSEXP, SEXP tolSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type y(ySEXP);
+    Rcpp::traits::input_parameter< int >::type dim(dimSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type W(WSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type D(DSEXP);
+    Rcpp::traits::input_parameter< int >::type iter(iterSEXP);
+    Rcpp::traits::input_parameter< double >::type tol(tolSEXP);
+    rcpp_result_gen = Rcpp::wrap(constrained_stress_major3D(y, dim, W, D, iter, tol));
+    return rcpp_result_gen;
+END_RCPP
+}
 // reweighting
 NumericVector reweighting(IntegerMatrix el, List N_ranks);
 RcppExport SEXP _graphlayouts_reweighting(SEXP elSEXP, SEXP N_ranksSEXP) {
@@ -123,16 +152,48 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// stress3D
+double stress3D(NumericMatrix x, NumericMatrix W, NumericMatrix D);
+RcppExport SEXP _graphlayouts_stress3D(SEXP xSEXP, SEXP WSEXP, SEXP DSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type W(WSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type D(DSEXP);
+    rcpp_result_gen = Rcpp::wrap(stress3D(x, W, D));
+    return rcpp_result_gen;
+END_RCPP
+}
+// stress_major3D
+NumericMatrix stress_major3D(NumericMatrix y, NumericMatrix W, NumericMatrix D, int iter, double tol);
+RcppExport SEXP _graphlayouts_stress_major3D(SEXP ySEXP, SEXP WSEXP, SEXP DSEXP, SEXP iterSEXP, SEXP tolSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type y(ySEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type W(WSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type D(DSEXP);
+    Rcpp::traits::input_parameter< int >::type iter(iterSEXP);
+    Rcpp::traits::input_parameter< double >::type tol(tolSEXP);
+    rcpp_result_gen = Rcpp::wrap(stress_major3D(y, W, D, iter, tol));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 static const R_CallMethodDef CallEntries[] = {
     {"_graphlayouts_constrained_stress", (DL_FUNC) &_graphlayouts_constrained_stress, 3},
     {"_graphlayouts_constrained_stress_major", (DL_FUNC) &_graphlayouts_constrained_stress_major, 6},
+    {"_graphlayouts_constrained_stress3D", (DL_FUNC) &_graphlayouts_constrained_stress3D, 3},
+    {"_graphlayouts_constrained_stress_major3D", (DL_FUNC) &_graphlayouts_constrained_stress_major3D, 6},
     {"_graphlayouts_reweighting", (DL_FUNC) &_graphlayouts_reweighting, 2},
     {"_graphlayouts_sparseStress", (DL_FUNC) &_graphlayouts_sparseStress, 6},
     {"_graphlayouts_stress", (DL_FUNC) &_graphlayouts_stress, 3},
     {"_graphlayouts_stress_major", (DL_FUNC) &_graphlayouts_stress_major, 5},
     {"_graphlayouts_stress_radii", (DL_FUNC) &_graphlayouts_stress_radii, 5},
     {"_graphlayouts_stress_focus", (DL_FUNC) &_graphlayouts_stress_focus, 7},
+    {"_graphlayouts_stress3D", (DL_FUNC) &_graphlayouts_stress3D, 3},
+    {"_graphlayouts_stress_major3D", (DL_FUNC) &_graphlayouts_stress_major3D, 5},
     {NULL, NULL, 0}
 };