+++ NEWS.md
@@ -1,4 +1,7 @@
-# graphlayouts (development version)
+# graphlayouts 0.7.0
+
+* added `layout_as_multilevel()` for multilevel networks
+* added `layout_with_stress3D()` and `layout_with_constrained_stress3D()` for 3D layouts
 
 # graphlayouts 0.6.0